Updated ReadMe file with code snippets
I added a code snippet to help set up card with a test card information and to check for validity of the card. Also, I added a code to  show how and where the Charge object is instantiated and the setCard() method called before performing charge. This is done to ensure that users find a sample code they can work with for a seamless integration.